Reduced processing of unvisited nodes in UCTNode::uct_select_child()
For unvisited nodes the score is proportional to the calculated value. Therefore for the unvisited nodes the best score can be found whilst counting the parent visits and used once to calculate the best value for those nodes. This then allows the second loop to only be processed if there are parent visits and only has to process nodes that have visits > 0.