Experiment/proof of concept for finding which NN input state elements have the most impact on the NN outputs.
This is done by iterating over all the input elements and flipping each one in turn, therefore generating output data for each "neighbouring" input state. For a specified NN output, the average difference to the current state NN output is output as a heatmap in terms of which input is flipped. In this example:
- the input flipping is dumbly going over each individual input. Therefore a vertex with a stone will be flipped in only one side's inputs (meaning it is both black and white). Also, no attempt is made to make the input state consistent over the history planes i.e. we will be inputting states where stones appear and disappear.
- history plane input changes are averaged together i.e. there is only one board output rather than one for each history plane.

GTP command syntax:
ndiff ["winrate"|"pass"|vertex]